// src/types.ts
export interface StaticFlight {
  id: string;
  airline: string;
  flightNumber: string;
  destination: string;
  city: string;
}